Split ConfigLeaf into interface and implementation
